webapprt/test/content/sample.html

<!DOCTYPE HTML>

<!--
  This file is the actual test.  It's a webapp installed by webapprt_sample.html
  and loaded into its iframe.  It's just a plain mochitest.
-->

<html>
  <head>
    <meta charset="utf-8">
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  </head>
  <body>
    <p id="display">
      This is the test webapp.
    </p>
    <div id="content" style="display: none"></div>
    <pre id="test">
      <script>

SimpleTest.waitForExplicitFinish();

ok(true, "true is true!");

var self = navigator.mozApps.getSelf();
self.onsuccess = function () {
  ok(true, "onsuccess should be called");
  ok(self.result, "result should be nonnull");
  ok(self.result.manifest, "manifest should be nonnull");
  is(self.result.manifest.name, "Sample Test Webapp", "manifest.name");
  SimpleTest.finish();
};
self.onerror = function () {
  ok(false, "onerror should not be called");
  SimpleTest.finish();
};

      </script>
    </pre>
  </body>
</html>
